+++ zenovis/include/zenovis/ObjectsManager.h
@@ -8,6 +8,8 @@
 #include <string>
 #include <memory>
 #include <map>
+#include <tinygltf/json.hpp>
+using Json = nlohmann::json;
 
 namespace zenovis {
 
@@ -18,6 +20,8 @@ struct ObjectsManager : zeno::disable_copy {
     std::unordered_map<std::string, std::shared_ptr<zeno::IObject>> cached_mesh;
     std::map<std::string, std::shared_ptr<zeno::IObject>> lightObjects;
     bool needUpdateLight = true;
+    std::string staticSceneTree;
+    std::string dynamicSceneTree;
 
     template <class T = void>
     auto pairs() const {
@@ -33,6 +37,7 @@ struct ObjectsManager : zeno::disable_copy {
     ~ObjectsManager();
     void clear_objects();
     bool load_objects(std::map<std::string, std::shared_ptr<zeno::IObject>> const &objs, std::string& runtype);
+    std::map<std::string, std::shared_ptr<zeno::IObject>> scene_tree_to_descriptor(std::map<std::string, std::shared_ptr<zeno::IObject>> const &objs);
 
     std::optional<zeno::IObject*> get(std::string nid);
 };
